+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 ### Features
 
 - Add main thread ID to profiling payload (#1918)
+- Add App Hangs tracking (#1906)
+
+### Fixes
+
+- Remove WebKit optimization check (#1921)
 
 ## 7.18.1
 
@@ -15,7 +20,7 @@
 ## 7.18.0
 
 ### Features
-
+ 
 - Replace tracestate header with baggage (#1867)
 
 ### Fixes